requirements.txt
anthropic>=0.89.0
requests>=2.32.0
beautifulsoup4>=4.12.0
slack-sdk>=3.30.0
python-dotenv>=1.0.0
lxml>=5.0.0
